notification/serializers.py
from rest_framework import serializers
from .models import Alert

class AlertSerializer(serializers.ModelSerializer):
    class Meta:
        model = Alert
        fields = ['id', 'culture', 'timestamp', 'message', 'is_resolved']
