requirements.txt
Django==3.2.19
djangorestframework==3.14.0
django-allauth==0.55.2
django-cors-headers==4.2.0
djangorestframework-simplejwt==4.7.2
